output/requirements.txt
python-pptx>=0.6.23
Pillow>=10.0.0
imageio-ffmpeg>=0.4.9
